fix: refactor car to be independent of blockstore
