+++ tomb/src/types/blockstore/carv2.rs
@@ -9,12 +9,13 @@ use std::{
     path::{Path, PathBuf},
 };
 use tomb_common::{
-        blockstore::{
-            car::{
+    car::{
                 Streamable,
                 v1::block::Block,
                 v2::{index::indexable::Indexable, CarV2},
             },
+        blockstore::{
+            
             TombBlockStore,
         },
     utils::io::{get_read, get_read_write, get_write},
